Refactoring and iterating on making layout changes persist
